claude-conversation-vectordb/requirements.txt
anthropic>=0.18.0
chromadb>=0.4.22
sentence-transformers>=2.3.1
python-dotenv>=1.0.0
mcp>=0.1.0
pydantic>=2.0.0
tqdm>=4.66.0
